refactor(runner): tools.rs から汎用コマンド実行基盤を runner.rs へ抽出 (#66)
god module 化した tools.rs から汎用コマンド実行基盤を新規 src/runner.rs
へ逐語移設し、ゲートアダプタと分離する。

- runner.rs: GATE_TIMEOUT/MAX_OUTPUT_LINES/GateOutcome/ToolResult/
  kill_process_group/run_command/run_command_with_label + runner テスト8件
- tools.rs: ゲートアダプタを残置し crate::runner を pub(crate) 経由で消費
- reporter.rs/main.rs: 結果型参照を runner 直接参照へ切替え、
  reporter→tools 依存を切断 (tools 参照 0 件)

純移動で挙動不変。237 tests green、clippy --all-targets clean、fmt clean。

Closes #66
